Renamed module to uvp.snp
